requirements.txt
# ===========================================
# AI-crew Python Dependencies
# ===========================================

# --- LangGraph & LangChain ---
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0

# --- Aegra Server ---
# Note: Install Aegra from GitHub
# pip install git+https://github.com/ibbybuilds/aegra.git

# --- Database ---
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
sqlalchemy[asyncio]>=2.0.0
alembic>=1.13.0

# --- Web Framework ---
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# --- HTTP Client ---
httpx>=0.27.0
aiohttp>=3.9.0

# --- Observability ---
langfuse>=2.0.0

# --- GitHub Integration ---
PyGithub>=2.1.0

# --- Utilities ---
python-dotenv>=1.0.0
pyyaml>=6.0.0
tenacity>=8.2.0
typing-extensions>=4.7.0

# --- Development ---
pytest>=8.0.0
pytest-asyncio>=0.23.0
black>=24.0.0
ruff>=0.4.0
mypy>=1.9.0
